Further restructuring and rebalancing
